athena/athena.py
import boto3

def batch_get_named_query():
    athena = boto3.client('athena')
    response = athena.batch_get_named_query(
        NamedQueryIds=[
            'string'
        ]
    )
    return response

def batch_get_query_execution():
    athena = boto3.client('athena')
    response = athena.batch_get_query_execution(
        QueryExecutionIds=[
            'string'
        ]
    )
    return response

def can_paginate():
    athena = boto3.client('athena')
    response = athena.can_paginate("foo_bar_operation")
    return response

def create_named_query():
    athena = boto3.client('athena')
    response = athena.create_named_query(
        Name='string',
        Description='string',
        Database='string'
    )
    return response

def create_work_group():
    athena = boto3.client('athena')
    response = athena.create_work_group(
        Name='string'
    )
    return response

def delete_data_catalog():
    athena = boto3.client('athena')
    response = athena.delete_data_catalog(
        Name='string'
    )
    return response

def delete_work_group():
    athena = boto3.client('athena')
    response = athena.delete_work_group(
        Name='string'
    )
    return response

def generate_presigned_url():
    athena = boto3.client('athena')
    response = athena.generate_presigned_url(
        ClientMethod='foo_bar_method'
    )
    return response

def get_data_catalog():
    athena = boto3.client('athena')
    response = athena.get_data_catalog(
        Name='string'
    )
    return response

def get_database():
    athena = boto3.client('athena')
    response = athena.get_database(
        Name='string'
    )
    return response

def get_named_query():
    athena = boto3.client('athena')
    response = athena.get_named_query(
        NamedQueryId='string'
    )
    return response

# TODO: Implement
# def get_paginator():
#     athena = boto3.client('athena')
#     response = athena.get_paginator(
#         get_paginator="foo_bar_operation"
#     )
#     return response

def get_query_execution():
    athena = boto3.client('athena')
    response = athena.get_query_execution(
        QueryExecutionId='string'
    )
    return response

def get_query_results():
    athena = boto3.client('athena')
    response = athena.get_query_results(
        QueryExecutionId='string'
    )
    return response

def get_table_metadata():
    athena = boto3.client('athena')
    response = athena.get_table_metadata(
        CatalogName='string',
        DatabaseName='string',
        TableName='string'
    )
    return response

# def get_waiter():
#     athena = boto3.client('athena')
#     response = athena.get_waiter("waiter_name")
#     return response

def get_work_group():
    athena = boto3.client('athena')
    response = athena.get_work_group(
        WorkGroup='string'
    )
    return response

def list_data_catalogs():
    athena = boto3.client('athena')
    response = athena.list_data_catalogs()
    return response

def list_databases():
    athena = boto3.client('athena')
    response = athena.list_databases(
        CatalogName='string'
    )
    return response
